feat: extend standard lib
