+++ example/.tmpl/std/convert/double.tmpl
@@ -1,0 +1,16 @@
+
+@extern fn d_toi(double value) : int;
+@extern fn d_tof(double value) : float;
+@extern fn d_tos(double value) : string;
+
+export fn double.toint() : int {
+    return d_toi(self);
+}
+
+export fn double.tofloat() : float {
+    return d_tof(self);
+}
+
+export fn double.tostring() : string {
+    return d_tos(self);
+}
