+++ example/.tmpl/std/convert/int.c
@@ -1,8 +1,19 @@
 #include <stdlib.h>
 #include <stdio.h>
 #include <assert.h>
-#include <string.h>
 
+// --------- Integer -------------
+
+// Integer to Double
+void* i_tod(void** data, unsigned int argc) {
+    assert(argc == 1 && "Invalid argument amount.");
+    int* arg = (int*)data[0];
+    double* res = (double*)malloc(sizeof(double));
+    *res = *arg;
+    return (void*)res;
+}
+
+// Integer to Float
 void* i_tof(void** data, unsigned int argc) {
     assert(argc == 1 && "Invalid argument amount.");
     int* arg = (int*)data[0];
@@ -11,27 +22,11 @@ void* i_tof(void** data, unsigned int argc) {
     return (void*)res;
 }
 
-void* f_toi(void** data, unsigned int argc) {
-    assert(argc == 1 && "Invalid argument amount.");
-    float* arg = (float*)data[0];
-    int* res = (int*)malloc(sizeof(int));
-    *res = *arg;
-    return (void*)res;
-}
-
+// Integer to String
 void* i_tos(void** data, unsigned int argc) {
     assert(argc == 1 && "Invalid argument amount.");
     int* arg = (int*)data[0];
     char* res = (char*)malloc(sizeof(char));
     sprintf(res, "%ld", *arg);
-    return (void*)res;
-}
-
-void* cstrlen(void**data, unsigned int argc) {
-    assert(argc == 1 && "Invalid argument amount.");
-    char* str = (char*)data[0];
-    int len = strlen(str);
-    int* res = (int*)malloc(sizeof(int));
-    *res = len;
     return (void*)res;
 }
